🚧  add getblocks method to request client, add column block
I've added a method to the request client getBlocks() that will allow to query
multiple blocks at the same time from an array that is passed in. I have tied this
into the getContent() method of the Blocks so that they now return Blocks instead
of just an array of IDs, removing a step on the users side.